chore(workspace): clear pre-existing clippy and rustdoc lints
Work Context:
  - feat/content-ergonomics: the workspace-level clippy and rustdoc
    gates were already failing on main against several files unrelated
    to the content-ergonomics work. Cleared every error so the whole
    workspace passes `cargo clippy --workspace --all-targets -- -D
    warnings` and `RUSTDOCFLAGS=-D warnings cargo doc --workspace
    --no-deps`. No behavior changes — every fix is either a lint
    cleanup or a doc-link rewrite.

Changes:
* crates/tui-vfx-compositor/src/pipeline/cls_prepared_sampler.rs (Version 1.2.1):
  - Drop two redundant `as f32` casts on Gravity::new arguments
    (clippy::unnecessary_cast). SignalOrFloat::evaluate already
    returns f32.
* crates/tui-vfx-compositor/src/filters/cls_charset_noise.rs (Version 1.0.1):
  - Use struct-init form in the test make_cell helper instead of
    Default::default() followed by field assignment
    (clippy::field_reassign_with_default).
* crates/tui-vfx-compositor/tests/pipeline/test_orc_render_pipeline.rs (Version 5.4.1):
  - Drop a redundant ShadowConfig clone() since the type is Copy
    (clippy::clone_on_copy).
* crates/tui-vfx-style/src/models/cls_color_config.rs (Version 1.4.1):
  - Drop the broken intra-doc link to ratatui::style::Color (the style
    crate doesn't depend on ratatui directly). Use plain code
    formatting instead.
* crates/tui-vfx-style/src/models/cls_style_config.rs (Version 1.0.2):
  - Same fix for ratatui::style::Style.
* crates/tui-vfx-style/src/models/cls_spatial_shader_type.rs (Version 2.1.2):
  - Drop the broken intra-doc link to CompositionOptions::shader_layers
    (lives in the compositor crate, which the style crate doesn't
    depend on). Switch the StyleEffect::Spatial link to the full
    `crate::models::StyleEffect::Spatial` path so rustdoc resolves it.
* xtask/src/docs/extract_rustdoc.rs (Version 2.0.5):
  - Wrap the `effects["masks"]["Wipe"]` example in a code block so
    rustdoc stops parsing the bracketed string literal as an
    intra-doc link to "Wipe".

+++ crates/tui-vfx-compositor/src/pipeline/cls_prepared_sampler.rs
@@ -1,7 +1,7 @@
 // <FILE>tui-vfx-compositor/src/pipeline/cls_prepared_sampler.rs</FILE> - <DESC>Prepared sampler enum for pipeline rendering</DESC>
-// <VERS>VERSION: 1.2.0</VERS>
-// <WCTX>Add pendulum sampler for swaying animations</WCTX>
-// <CLOG>Add Pendulum variant to PreparedSampler</CLOG>
+// <VERS>VERSION: 1.2.1</VERS>
+// <WCTX>feat/content-ergonomics: clean up pre-existing workspace clippy lint</WCTX>
+// <CLOG>Drop redundant f32-to-f32 casts on Gravity::new arguments (clippy::unnecessary_cast)</CLOG>
 
 use crate::samplers::cls_bounce::Bounce;
 use crate::samplers::cls_crt_jitter::CrtJitter;
@@ -193,10 +193,10 @@ pub(crate) fn prepare_sampler(t: f64, sampler_spec: &Option<SamplerSpec>) -> Pre
         } => {
             let eval_accel = acceleration.evaluate(t, &signal_ctx).unwrap_or(4.0);
             let eval_terminal = terminal_velocity.evaluate(t, &signal_ctx).unwrap_or(10.0);
-            PreparedSampler::Gravity(Gravity::new(eval_accel as f32, eval_terminal as f32, *axis))
+            PreparedSampler::Gravity(Gravity::new(eval_accel, eval_terminal, *axis))
         }
     }
 }
 
 // <FILE>tui-vfx-compositor/src/pipeline/cls_prepared_sampler.rs</FILE> - <DESC>Prepared sampler enum for pipeline rendering</DESC>
-// <VERS>END OF VERSION: 1.2.0</VERS>
+// <VERS>END OF VERSION: 1.2.1</VERS>
